Move distribution abstraction in shared crate
