+++ crates/distribution-filename/src/lib.rs
@@ -1,7 +1,7 @@
 pub use source_distribution::{
     SourceDistributionExtension, SourceDistributionFilename, SourceDistributionFilenameError,
 };
-pub use wheel_filename::{WheelFilename, WheelFilenameError};
+pub use wheel::{WheelFilename, WheelFilenameError};
 
 mod source_distribution;
-mod wheel_filename;
+mod wheel;
