+++ crates/puffin-distribution/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "distribution-filename"
-version = "0.0.1"
+name = "puffin-distribution"
+version = "0.1.0"
 edition = { workspace = true }
 rust-version = { workspace = true }
 homepage = { workspace = true }
@@ -10,8 +10,8 @@ authors = { workspace = true }
 license = { workspace = true }
 
 [dependencies]
+distribution-filename = { path = "../distribution-filename" }
 pep440_rs = { path = "../pep440-rs" }
-platform-tags = { path = "../platform-tags" }
 puffin-package = { path = "../puffin-package" }
 
-thiserror = { workspace = true }
+anyhow = { workspace = true }
