+++ crates/puffin-installer/Cargo.toml
@@ -14,6 +14,7 @@ install-wheel-rs = { path = "../install-wheel-rs", default-features = false }
 pep440_rs = { path = "../pep440-rs" }
 pep508_rs = { path = "../pep508-rs" }
 puffin-client = { path = "../puffin-client" }
+puffin-distribution = { path = "../puffin-distribution" }
 puffin-interpreter = { path = "../puffin-interpreter" }
 puffin-package = { path = "../puffin-package" }
 distribution-filename = { path = "../distribution-filename" }
