+++ crates/puffin-resolver/Cargo.toml
@@ -17,6 +17,7 @@ platform-host = { path = "../platform-host" }
 platform-tags = { path = "../platform-tags" }
 pubgrub = { path = "../../vendor/pubgrub" }
 puffin-client = { path = "../puffin-client" }
+puffin-distribution = { path = "../puffin-distribution" }
 puffin-package = { path = "../puffin-package" }
 puffin-traits = { path = "../puffin-traits" }
 distribution-filename = { path = "../distribution-filename" }
